mm/huge_memory: use new hashtable implementation
Switch hugemem to use the new hashtable implementation. This reduces the amount of
generic unrelated code in the hugemem.

This also removes the dymanic allocation of the hash table. The size of the table is
constant so there's no point in paying the price of an extra dereference when accessing
it.

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>

Conflicts:
	mm/huge_memory.c

Former-commit-id: a050a98786d5557399fcdd0643fe61c2eef08e7b